Changed use of command multiplier.
The multiplier is now entered after the command indicating how often the command should be reapplied.

This is both cleaner in implementation and in the cognitive model for the user:
'delete... oh twice' instead of: 'do 2 times delete' meaning that the action
can unfold while performing instead of having to be defined before execution.

Besides, now the dot operator is replaced by the same command as the repeat operator